refactor: extract router support classes out of app_router
app_router.dart held several classes beyond the route table. Move each
to its own file so the router file is just buildRouter plus its small
route-builder helpers:

- ActiveMatrixListenable  -> active_matrix_listenable.dart
- AccountSwitchRedirector -> account_switch_redirector.dart
- AddAccountShell         -> widgets/add_account_shell.dart
- AdaptiveCallScreen      -> features/calling/screens/adaptive_call_screen.dart

The two private widgets are now public so they can live in their own
files. Also add a RoutePaths.roomPrefix constant for the redirector's
room-route check. app_router.dart drops from 455 to 333 lines.
